warning fix

+++ flag_funcs2.c
@@ -60,10 +60,10 @@ void flag_n(int *i, int nb)
     *i = nb;
 }
 
-int flag_p(unsigned int *i)
+unsigned int flag_p(int *i)
 {
     my_putstr("0x7fff");
-    return flag_x((unsigned int)i);
+    return flag_x(i);
 }
 
 double is_neg(double nb, int *nb_of_char)
